requirements.txt
phonenumbers==8.13.2
pymysql==1.1.2
python-dotenv==1.1.1
configparser==7.2.0
tqdm==4.67.1
pandas==2.2.3
numpy==2.1.1